chore: improve overall styles in merchant admin
