+++ packages/merchant-admin/styles/button.scss
@@ -10,8 +10,8 @@
   padding: 0.5rem 1rem;
 
   &:hover {
-    background-color: var(--gray-300);
-    color: var(--text-color);
+    background-color: var(--text-color);
+    color: #fff;
   }
 
   &[disabled="disabled"],
