+++ packages/merchant-admin/styles/form.scss
@@ -1,34 +1,55 @@
 
 .form {
+  padding: 2rem;
+  background-color: var(--gray-100);
+  border: 1px solid var(--gray-300);
+  border-radius: 3px;
+  color: var(--text-color);
+
   button[type="submit"] {
     margin-top: 1rem;
     font-size: 1.2rem;
-
-    &:hover {
-      background-color: var(--green-700) !important;
-    }
+    padding: 0.75rem 2rem;
   }
 
   label {
     text-transform: uppercase;
-    font-weight: 500;
-    color: var(--gray-600);
+    font-weight: 600;
+    color: var(--text-color);
+    opacity: 0.9;
   }
 
-  color: var(--text-color);
+  &__input {
+    display: block;
+    width: 100%;
+    margin-top: 0.5rem;
+    margin-bottom: 1rem;
+    padding: 0.75rem 0.75rem;
+    font-size: 1.2rem;
+    border: 1px solid var(--gray-500);
+    border-radius: 0px;
+    color: var(--text-color);
+    border: 1px solid var(--gray-300);
+  }
+
+  &__field {
+    margin-bottom: 2rem;
+
+    &:first-child {
+      margin-top: 2rem;
+    }
+  }
+
+  &__title {
+    font-size: 1.25rem;
+    font-weight: 500;
+    margin-bottom: 2.5rem;
+  }
+
+  &__description {
+    font-size: 1rem;
+    margin-bottom: 2.5rem;
+    margin-top: -1.5rem;
+  }
 }
 
-.field {
-  margin-bottom: 2rem;
-}
-
-.form-input {
-  display: block;
-  width: 100%;
-  margin-top: 0.5rem;
-  margin-bottom: 1rem;
-  padding: 0.5rem 0.75rem;
-  font-size: 1.2rem;
-  border: 1px solid var(--gray-500);
-  border-radius: 3px;
-}
