+++ packages/merchant-admin/styles/list.scss
@@ -1,40 +1,57 @@
 .list {
   display: flex;
   flex-direction: column;
+  position: relative;
 
   &__header {
-    padding: 0rem 1rem;
-    margin-bottom: 0;
+    padding: 0rem 1.5rem;
+    margin-bottom: 0 !important;
   }
 
   &__title {
     text-transform: uppercase;
-    font-weight: 500;
     color: var(--gray-500);
     padding-left: 0;
+    font-weight: 600;
+    font-size: 0.9rem;
   }
 }
 
 .list-item {
   padding: 1.5rem;
+  display: block;
   background-color: var(--gray-100);
   border: 1px solid var(--gray-300);
   border-radius: 3px;
-  margin-bottom: 1rem;
+  margin-bottom: 0.75rem;
   align-items: center;
   font-size: 0.9rem;
-
-  // .btn-delete-app {
-  //   display: block;
-  //   visibility: hidden;
-  //   margin-left: 1rem;
-  //   transform: rotateZ(45deg);
-  //   text-align: right;
-  //   font-size: 1.5rem;
-  // }
+  position: relative;
+  overflow: hidden;
 
   &--clickable {
     cursor: pointer;
+  }
+
+  &--loading {
+    &::after {
+      position: absolute;
+      top: 0;
+      right: 0;
+      bottom: 0;
+      left: 0;
+      transform: translateX(-100%);
+      background-image: linear-gradient(
+        90deg,
+        rgba(#edf2f7, 0) 0,
+        rgba(#edf2f7, 0.25) 25%,
+        rgba(#edf2f7, 0.5) 50%,
+        rgba(#edf2f7, 0.75) 75%,
+        rgba(#edf2f7, 0)
+      );
+      animation: shimmer 1s infinite;
+      content: "";
+    }
   }
 
   &:hover {
