+++ packages/merchant-admin/styles/main.scss
@@ -3,8 +3,6 @@
 @import "../../../node_modules/bulma/sass/helpers/flexbox.sass";
 @import "../../../node_modules/bulma/sass/utilities/_all.sass";
 @import "../../../node_modules/bulma/sass/grid/_all.sass";
-@import "../../../node_modules/bulma/sass/elements/table.sass";
-@import "../../../node_modules/bulma/sass/elements/form.sass";
 
 @import "./animations.scss";
 @import "./button.scss";
@@ -12,6 +10,8 @@
 @import "./metric-box.scss";
 @import "./modal.scss";
 @import "./list.scss";
+@import "./icons.css";
+@import "./sidebar.scss";
 
 $header_height: 4.5rem;
 
@@ -30,7 +30,7 @@ $header_height: 4.5rem;
   --green-100: #c6f6d5;
   --green-200: #9ae6b4;
   --green-300: #48bb78;
-  --text-color: #363636;
+  --text-color: #2d3748;
   --blue-100: #BEE3F8;
   --blue-200: #90CDF4;
   --blue-300: #63B3ED;
@@ -50,11 +50,12 @@ body {
 }
 
 body {
-  font-family: -apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,Open Sans,Helvetica Neue,sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,Open Sans,Helvetica Neue,sans-serif;
   color: var(--text-color);
   margin: 0;
   padding: 0;
   width: 100%;
+  overflow: hidden;
 }
 
 p {
@@ -64,23 +65,6 @@ p {
 #root {
   height: 100%;
   width: 100%;
-}
-
-.logo {
-  font-size: 1.4rem;
-  font-weight: 600;
-  text-align: left;
-
-  a {
-    color: var(--text-color);
-  }
-
-  &:hover {
-    a {
-      color: var(--text-color);
-      text-decoration: underline;
-    }
-  }
 }
 
 .main {
@@ -94,31 +78,7 @@ p {
   max-width: 1280px;
   width: 100%;
   margin: 0 auto;
-}
-
-.sidebar {
-  padding: 2rem;
-  width: 300px;
-  background-color: var(--gray-100);
-  border-right: 1px solid var(--border-color);
-  height: 100%;
-}
-
-.sidebar-menu {
-  margin-top: 2rem;
-  display: flex;
-  flex-direction: column;
-}
-
-.sidebar-menu-item {
-  padding-top: 0.5rem;
-  font-size: 1.2rem;
-  font-weight: 500;
-  color: var(--blue-700) !important;
-
-  &:hover {
-    text-decoration: underline;
-  }
+  overflow-y: auto;
 }
 
 .link {
@@ -147,7 +107,7 @@ p {
   display: flex;
   flex-direction: row;
   justify-content: space-between;
-  align-items: center;
+  align-items: flex-start;
 }
 
 .login-page {
@@ -180,14 +140,6 @@ p {
   &:hover {
     opacity: 1;
   }
-}
-
-.new-product-form {
-  max-width: 700px;
-  background-color: var(--gray-100);
-  border: 1px solid var(--border-color);
-  border-radius: 3px;
-  padding: 2rem;
 }
 
 .icon-button {
